requirements.txt
pyjwt==2.8.0
flask==3.0.3
flask-cors==4.0.0
bcrypt==4.1.2